+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## []
+
+- Improved auto-detected tooltip
+
 ## [1.8.2]
 
 - Fixed sporadic game freeze on world start